+++ flashback/formatting/__init__.py
@@ -1,5 +1,6 @@
 from .functions import adverbize
 from .functions import camelize
+from .functions import kebabize
 from .functions import ordinalize
 from .functions import oxford_join
 from .functions import parameterize
@@ -13,6 +14,7 @@ from .functions import truncate
 __all__ = (
     'adverbize',
     'camelize',
+    'kebabize',
     'oxford_join',
     'ordinalize',
     'parameterize',